refactor: SchoolCreateCommand, SchoolUpdateCommand 패키지 이동 및 검증 로직 제거
- 도메인 응집도 향상을 위해 DTO 검증 로직 제거

(cherry picked from commit ab06daaff355c56ac7e68d034f33333603eb814c)
