+++ backend/src/main/java/com/festago/school/dto/command/SchoolUpdateCommand.java
@@ -1,0 +1,14 @@
+package com.festago.school.dto.command;
+
+import com.festago.school.domain.SchoolRegion;
+import lombok.Builder;
+
+@Builder
+public record SchoolUpdateCommand(
+    String name,
+    String domain,
+    SchoolRegion region,
+    String logoUrl,
+    String backgroundImageUrl
+) {
+}
