+++ backend/src/test/java/com/festago/school/application/SchoolCommandServiceTest.java
@@ -10,8 +10,8 @@ import com.festago.common.exception.ErrorCode;
 import com.festago.common.exception.NotFoundException;
 import com.festago.school.domain.School;
 import com.festago.school.domain.SchoolRegion;
-import com.festago.school.dto.SchoolCreateCommand;
-import com.festago.school.dto.SchoolUpdateCommand;
+import com.festago.school.dto.command.SchoolCreateCommand;
+import com.festago.school.dto.command.SchoolUpdateCommand;
 import com.festago.school.repository.MemorySchoolRepository;
 import com.festago.school.repository.SchoolRepository;
 import com.festago.support.fixture.SchoolFixture;
